# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tsundoku API",
    "description": "Manga discovery service. Read endpoints (`GET`) are public unless `auth.read_requires_auth = true`. Write endpoints (`POST`) always require the admin bearer.",
    "contact": {
      "name": "Sylvain Cau",
      "email": "sylvain.cau@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/series": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "List series ordered by last release timestamp (most recent first by default).",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "kind",
            "in": "query",
            "description": "Filter by stored `series.type` (e.g. `manga`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by stored `series.status` (e.g. `ongoing`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owned",
            "in": "query",
            "description": "Filter by ownership flag (true = owned by Codex, false = discoverable).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field. Supports `last_release_at` (default) and `first_seen_at`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "`asc` or `desc` (default).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesListPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/series/{id}": {
      "get": {
        "tags": [
          "series"
        ],
        "summary": "Series detail, including the resolved external-ID mappings.",
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Internal series id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "404": {
            "description": "No series with that id"
          }
        }
      }
    },
    "/api/v1/series/{id}/refresh-metadata": {
      "post": {
        "tags": [
          "series"
        ],
        "summary": "Re-fetch metadata for a series from the active provider and re-persist.",
        "operationId": "refresh_metadata",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Internal series id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetail"
                }
              }
            }
          },
          "404": {
            "description": "Series or provider entry not found"
          },
          "409": {
            "description": "No mapping for the active provider on this series"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
...
    "/api/v1/sources/{name}/poll": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Trigger a one-shot poll for the named source. Uses the same per-source\nmutex the cron job holds, so a manual kick during a scheduled tick is\nsilently skipped (`skipped = true`).",
        "operationId": "poll",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Source instance name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualPollResponse"
                }
              }
            }
          },
          "404": {
            "description": "No source with that name registered"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExternalIdDto": {
        "type": "object",
        "required": [
          "provider",
          "externalId",
          "fetchedAt"
        ],
        "properties": {
          "externalId": {
            "type": "string"
          },
          "externalUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "fetchedAt": {
            "type": "integer",
            "format": "int64"
          },
          "provider": {
            "type": "string"
          }
        }
      },
...
      "ManualPollResponse": {
        "type": "object",
        "required": [
          "source",
          "triggered",
          "skipped"
        ],
        "properties": {
          "skipped": {
            "type": "boolean",
            "description": "`false` when a previous tick is still in flight; the request is a no-op."
          },
          "source": {
            "type": "string"
          },
          "triggered": {
            "type": "boolean"
          }
        }
      },
...
      "SeriesDetail": {
        "type": "object",
        "required": [
          "id",
          "canonicalTitle",
          "alternateTitles",
          "genres",
          "metadataSource",
          "metadataFetchedAt",
          "firstSeenAt",
          "lastReleaseAt",
          "owned",
          "externalIds"
        ],
        "properties": {
          "alternateTitles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "canonicalTitle": {
            "type": "string"
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "externalIds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalIdDto"
            }
          },
          "firstSeenAt": {
            "type": "integer",
            "format": "int64"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "highestChapter": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "highestVolume": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastReleaseAt": {
            "type": "integer",
            "format": "int64"
          },
          "metadataFetchedAt": {
            "type": "integer",
            "format": "int64"
          },
          "metadataSource": {
            "type": "string"
          },
          "owned": {
            "type": "boolean"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "SeriesListItem": {
        "type": "object",
        "required": [
          "id",
          "canonicalTitle",
          "lastReleaseAt",
          "firstSeenAt",
          "owned"
        ],
        "properties": {
          "canonicalTitle": {
            "type": "string"
          },
          "coverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "firstSeenAt": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastReleaseAt": {
            "type": "integer",
            "format": "int64"
          },
          "owned": {
            "type": "boolean"
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
      "SeriesListPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesListItem"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "admin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Admin bearer token (auth.admin_token in config)."
      }
    }
  },
...
}